Refactor, PEP8 Fixes, script "codemod", release PyPI
