+++ .travis.yml
@@ -3,6 +3,8 @@ language: python
 python:
   - "2.7"
 
-install: "pip install ."
-script: "codemod.py --test"
+install:
+  - "pip install ."
+  - "pip install flake8 pytest"
+script: make test
