+++ README.md
@@ -7,11 +7,11 @@ Status](https://travis-ci.org/facebook/codemod.svg?branch=master)](https://travi
 Overview
 --------
 
-codemod.py is a tool/library to assist you with large-scale codebase refactors that can be partially automated but still require human oversight and occassional intervention.
+codemod is a tool/library to assist you with large-scale codebase refactors that can be partially automated but still require human oversight and occassional intervention.
 
 Example: Let's say you're deprecating your use of the `<font>` tag.  From the command line, you might make progress by running:
 
-    codemod.py -m -d /home/jrosenstein/www --extensions php,html \
+    codemod -m -d /home/jrosenstein/www --extensions php,html \
         '<font *color="?(.*?)"?>(.*?)</font>' \
         '<span style="color: \1;">\2</span>'
 
@@ -19,7 +19,7 @@ For each match of the regex, you'll be shown a colored diff, and asked if you wa
 
 Install
 -------
-`pip install git+https://github.com/facebook/codemod.git`
+`pip install codemod`
 
 Usage
 -----
