+++ codemod/__init__.py
@@ -1,0 +1,1 @@
+from codemod.base import *  # noqa
