+++ pytest.ini
@@ -1,0 +1,3 @@
+# content of pytest.ini
+[pytest]
+addopts = --doctest-modules codemod
